Bugfixing sca and removing duplicate functions
