+++ python/reg_interface/common/gpio.py
@@ -1,7 +1,7 @@
 from reg_xml_parser import getNode
 from reg_base_ops import writeReg
 from bit_utils import hex
-from sca_utils import getOHlist, sendScaCommand
+from sca_common_utils import getOHlist, sendScaCommand
 def set_direction(ohMask, directionMask):
     subheading('Disabling monitoring')
     writeReg(getNode('GEM_AMC.SLOW_CONTROL.SCA.ADC_MONITORING.MONITORING_OFF'), 0xffffffff)
